bug fix: qt widget layout error (#302)
This pull request makes minor adjustments to the initialization of UI
layouts in the `widgets.py` file to improve consistency and avoid
passing the parent widget unnecessarily. These changes simplify the
layout creation process.

### UI Layout Initialization Improvements:
*
[`software/control/widgets.py`](diffhunk://#diff-26413fa85447f35aa4077535ae2d5f6f1905d5de96d6b303d426f9d039ddde2bL994-R996):
Removed the parent widget argument (`self`) when initializing
`QVBoxLayout`, `QHBoxLayout`, and `QGridLayout` in the `init_ui` method.
This ensures a cleaner and more consistent approach to layout
initialization.
[[1]](diffhunk://#diff-26413fa85447f35aa4077535ae2d5f6f1905d5de96d6b303d426f9d039ddde2bL994-R996)
[[2]](diffhunk://#diff-26413fa85447f35aa4077535ae2d5f6f1905d5de96d6b303d426f9d039ddde2bL1012-R1012)